pyproject.toml
[tool.poetry]
name = "noe4j-prez"
version = "0.1.0"
description = ""
authors = ["tom <ttaranto@ekinox.io>"]
readme = "README.md"
packages = [{include = "noe4j_prez"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
neo4j = "^5.6.0"
numpy = "^1.24.2"
tqdm = "^4.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
